app/firstsaturdaybot/conversation_handlers/common_conversation.py
from firstsaturdaybot.commands.common_commads import *
from firstsaturdaybot.commands import *


from telegram.ext import (
    CommandHandler, 
    MessageHandler, 
    filters, 
    ConversationHandler, 
    CallbackQueryHandler)

unknown_handler = MessageHandler(filters.COMMAND, unknown_command)

end_handler = MessageHandler(filters.COMMAND, end_command)